third_party/blink/web_tests/fast/js/method-check-expected.txt

Test method-check related bugs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS total is 200
PASS successfullyParsed is true

TEST COMPLETE
